second commit

+++ src/calculadoraDigital/Calculadora.java
@@ -1,67 +1,34 @@
 package calculadoraDigital;
-
 public class Calculadora {
-
-
 
 // atributo 
 
-private int resultado = 0; 
-
-
+private double resultado = 0; 
 
 // método somar 
 
-public int somar( int n1, int n2 ){ 
-
-
-
+public double somar( int n1, int n2 ){ 
     resultado = n1 + n2; 
-
     return resultado; 
-
 } 
-
-
 
 // método subtrair 
-
-public int subtrair(int n1, int n2 ){ 
-
-
-
+public double subtrair(int n1, int n2 ){ 
     resultado = n1 - n2; 
-
     return resultado; 
-
 }         
 
-
-
 // método multiplicar 
-
-public int multiplicar( int n1, int n2 ){ 
-
-
-
+public double multiplicar( int n1, int n2 ){ 
+	
     resultado = n1 * n2; 
-
     return resultado; 
-
 } 
 
-
-
 // método dividir 
-
-public int dividir( int n1, int n2 ){ 
-
-
-
+public double dividir( int n1, int n2 ){ 
     resultado = n1 / n2; 
-
     return resultado; 
-
 } 
 
 }
